requirements.txt
flask==2.3.3
mcstatus==10.0.3
pysocks==1.7.1
stem==1.8.1
requests==2.31.0
